Added functionality for counting flows and groups.
This patch contains functionality for counting flows / groups.
The existing tests have been modified in order to work this way,
however, we use a flag that indicates if to check flows and groups.
As of now it is False, since I observed that between some iterations
the same test produced different flows and groups at the switches.

This might be a bug that we need to discuss about.

Change-Id: Ie56333623c6e0eb9d8e03f51894078bcd272df2e